requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
loguru>=0.7.0
bcrypt>=4.0.0
pdfplumber>=0.10.0
alembic>=1.13.0